[#21054] docdb: Add ABORTED state to clone state manager
Summary:
This diff adds the `ABORTED` state and an `abort_message` field to the `SysCloneStateInfoPB` object. A clone is moved to the `ABORTED` state if it fails some step of the cloning processes, or if it was incomplete (not `ABORTED` or `RESTORED`) upon a master failover. There is currently no timeout behavior for clones stuck in tablet creation or the restoring state.

Aborted clones do not currently clean up their namespace; this is left to the user. Aborted clones do not block clones to a different target namespace.

**Upgrade/Rollback safety:**
Only touches a non-production protobuf, guarded by `FLAGS_enable_db_clone`.

Jira: DB-10021

Test Plan:
```
./yb_build.sh --cxx-test clone_state_manager-test --gtest_filter CloneStateManagerTest.AbortInStartTabletsCloning
./yb_build.sh --cxx-test clone_state_manager-test --gtest_filter CloneStateManagerTest.AbortIfFailToSchedulePgCloneSchema
./yb_build.sh --cxx-test clone_state_manager-test --gtest_filter CloneStateManagerTest.AbortInPgSchemaClone
./yb_build.sh --cxx-test clone_state_manager-test --gtest_filter CloneStateManagerTest.AbortInStartTabletsCloningPg
./yb_build.sh --cxx-test clone_state_manager-test --gtest_filter CloneStateManagerTest.AbortInCreatingState
./yb_build.sh --cxx-test clone_state_manager-test --gtest_filter CloneStateManagerTest.AbortInRestoringState
./yb_build.sh --cxx-test clone_state_manager-test --gtest_filter CloneStateManagerTest.AbortIncompleteCloneOnLoad
```

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32999