[BACKPORT 2024.1][#21765][#21686][#21701][#21841][#21864] YSQL: Backport walsender fixes for PG compatible logical replication
Summary:
##### Backport Description
All cherry-picks were clean. No merge conflicts.

##### Original Description(s)

Original commits:
612880ed8405a10cbbd19a04fb9abb2fb5d83bf6 / D33750
c6bbdae93dbf274a9787f9ed2aab3643d6b4c9d3 / D33647
a4efd56d347962510faf724559f1b1007aa3ffde / D33649
f2a28210c62f6a24a62821e864c3e4a5a210aab3 / D33876
fa41967b4bd7269e9721bc2e283ccbf055365e5f / D33910

###### YSQL: Fix handling of serialized transactions in reorder buffer
During logical replication via the walsender, if a transaction has more than max_changes_in_memory (4096) DML records, it is serialized to the disk. They are
written to the `pg_repl_slot/{slot_name}/` directory. In PG, this directory is created as part of the slot creation and also stores the stream metadata and any
exported snapshots.

In YB, we had disabled the creation of this directory during slot creation because:
1. yb-master is the source of truth
2. One could create the slot on one node and then connect on another to consume changes.

This meant that when a transaction needs to be serialized, the reorder buffer would run into error since the directory doesn't exist.

This revision add the logic to create this directory upon the start of replication. This needs to happen at the start of streaming instead of slot creation due to the
point 2 above.

###### YSQL: Require replica identity to be enabled for replication slot creation and consumption
Replica identity support is required to stream UPDATE operations via the logical replication protocol. This revision adds validations at the time of slot
creation and consumption that it is enabled.

Also, minor cleanups in the code.

###### YSQL: Send the replica identity of the table which existed at the time of slot creation.
As part of the RELATION message, the replica identity of the relation is sent during logical replication.

In YSQL, the effective replica identity that is used is the value which existed at the time of slot creation. Future Alter Table Replica Identity statements only
affect streams created after it but not before it. As a result, we now send the correct effective replica identity as part of the RELATION message.

###### YSQL: Disable loading replication slots from disk on startup
In PG, the replication slot metadata, snapshots and spilled transactions are stored in the `pg_replslot/{slot_name}` directory. Upon server
startup, the metadata is read from the disk and loaded into shared memory.

In YB, we only utilize the disk for storing spilled transactions during streaming (https://phorge.dev.yugabyte.com/D33750). This revision disables loading the replication slots from the disk as the metadata doesn't exist so it leads to failures such as #21841. Note that in YB, the replication slot metadata is stored in yb-master and loaded on demand when asked to stream.

###### YSQL: Compute and log the time taken in different parts of the streaming protocol in both walsender and Virtual WAL
We do not have sufficient means to track the amount of time taken by the different stages in the logical replication streaming protocol.

This revision introduces the computation and logging of these times aiming at improving the debugging in case of performance issues.
Jira: DB-10640, DB-10570, DB-10581, DB-10741, DB-10762

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*

```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testReplicationWithSerializedTransaction'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testWithAlterReplicaIdentity'
```

Reviewers: asrinivasan, skumar

Reviewed By: asrinivasan

Subscribers: yql, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D34325