[#20676] YSQL, ASH: Update default query_id in ASH metadata
Summary:
This diff sets a default query_id of `5` to ASH metadata so that catalog requests
are not ignored and we can know the distribution of time spent between catalog
and storage requests. The root_request_id is still empty for these catalog requests,
as there is no point in having different root_request_id for different catalog requests now.

In future, we would want to have the actual query_id for query catalog requests and
only the catalog requests that are not tied to any specific query can have this query_id
of `5`.

This diff also gets rid of the field `yb_is_ash_metadata_set` since now every outgoing
Perform RPC will have a default query_id. But this causes another problem, an idle ysql
session will continue to fill the circular buffer with `ClientRead` wait events, which is not
ideal, because this might overwrite actual useful information. We filter these out so
that it doesn't eat up space in the circular buffer.
Jira: DB-9677

Test Plan: ./yb_build.sh --java-test TestYbAsh#testCatalogRequests

Reviewers: jason

Reviewed By: jason

Subscribers: amitanand, hbhanawat, yql

Differential Revision: https://phorge.dev.yugabyte.com/D33912