[BACKPORT 2024.1][PLAT-13500] Fallback to YBA's default AMI for YBA managed bundles
Summary:
In case the AMI is not present for a region in bundle, we were falling
back to retrieve the same from the region.
This diff makes changes as -
1. For YBA managed bundles - we will read the AMIs from the YBA metadata in
case they are not present in the bundle.

2. For Custom bundles - We will fail without having any fall back mechanism.

Removes the dependency on region -> ybImage.
Depends on - D34119
Original Commit - https://github.com/yugabyte/yugabyte-db/commit/bc3f1bd3f5fa27e65af28117caf223f5d7604360

Test Plan:
Created a provider with custom bundle.
Removed the ybImage from the bundle.
Deployed the universe using the same. Verified that it failed.

Created a provider with YBA managed bundles.
Removed the ybImage from the bundle.
Deployed the universe. Verified that it picks up from the YBA's image metadata.

Reviewers: amalyshev, nbhatia

Reviewed By: amalyshev

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D34265