[#22211] xCluster: Create new DB Scoped stream in INITIATED state
Summary:
Streams should be in INITIATED state until replication starts. This is done in order to block tablet splitting while we are in the setup phase.
The backup restore which sets up the tablets on the target cluster. When xCluster setup expects the tablet count to be the same to simplify the process.
Jira: DB-11131

Test Plan: Jenkins

Reviewers: jhe, xCluster

Reviewed By: jhe

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D34650