[#tbd] test: fix test logic in TestPgDepend#testSequenceDeletionWithCascade
Summary:
According to the test comments, the intention is to check that after a sequence that is used as a default value for a column is dropped, the dependency between the "table column type" and the sequence does not exist. However, such a dependency is never created and the logic for retrieving the dependency is incorrect:
```
      rs = statement.executeQuery("SELECT objid FROM pg_depend WHERE " +
                                      "classid=" + PG_TYPE_OID + " AND refobjid=" + oidSequence);
```
This code retrieves the dependency between the sequence and it's associated type.

This diff changes the test to instead check for the dependency between the column's default (metadata stored in `pg_attrdef`) and the sequence. Specifically, it changes the above query to:

```
      rs = statement.executeQuery("SELECT objid FROM pg_depend WHERE " +
          "classid=" + PG_ATTRDEF_OID + " AND refobjid=" + oidSequence);
```

Test Plan:
Jenkins: test regex: .*TestPgDepend.*

./yb_build.sh --java-test 'org.yb.pgsql.TestPgDepend#testSequenceDeletionWithCascade'

Reviewers: myang

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D34708