[DB-11170][xCluster]  Inconsistent Error Reporting for Missing DB IDs while making call against the YBA host
Summary:
Previously, YBA compared the DB name with the DB ids to find the missing DBs on the source universe while creating a DR config.

This diff makes changes to compare the DB ids and properly returns the list of missing DB ids.

Test Plan: - Make a request to create a dr config with two DB ids while one of the ids did not exist and made sure the correct error is returns to the user.

Reviewers: #yba-api-review!, cwang, vbansal, jmak, sanketh

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D34741