[PLAT-13720] legacy release api no longer appends release tag
Summary:
When the release redesign is enabled and the old release api is called, we no longer append the release tag to the end of the release version.

In addition, marked the old apis as deprecated, and fixed a small startup bug that was
causing app init to take a long time. We will no longer
consider helm charts as local files when attempting to import metadata, as helm charts
do not have that metadata

Test Plan:
validated that helm charts do not make a release "local"
validated old release api does not return release tag

Reviewers: muthu, anijhawan

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D34752