[PLAT-13808] Pass machineImage for disk update operation from the java layer
Summary:
Passes the machineImage if present in the nodeDetails as the args for disk update operation.
The order is -
1. use the machineImage present in the nodeDetails.
2. Fallback to the AMI present in the bundle.
3. Use the AMI id present in the host.

Test Plan:
Created a universe using machineImage on the fly.
Deleted the AMI from the console.
Performed disk update - verified that it fails.

Performed VM Image Upgrade.
Peformed disk update - Verified that it passes.

Reviewers: muthu, daniel, yshchetinin

Reviewed By: daniel

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D34811