[#21946] YSQL: Serialize yb_is_omitted array as part of the serialized txn in reorder buffer
Summary:
While streaming changes as part of the logical replication protocol, the reorderbuffer serializes large transactions to disk to avoid using excessive memory.

Apart from the datum and isnull values used to describe a heap tuple, we also have special handling for intentionally omitted values due to the CHANGE replica
identity. This array was not getting serialized as part of the serialization of the heap tuples, leading to the values being lost after restoring from the disk.

This revision updates the serialization and de-serialization logic to handle the yb_is_omitted array as well.
Jira: DB-10863

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*

./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot'

Reviewers: asrinivasan

Reviewed By: asrinivasan

Subscribers: yql, ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D34851