[PLAT-13520] Create alert for running out of tablets
Summary:
Create alert for running out of tablets.
DB team is introducing a new feature, the tablet guardrail mechanism, that will be available in 2024.1 onwards.

If the below two metrics exist (they are not present in older versions) and the first metric is not -1 then for each TServer the ratio `ts_live_tablet_peers/ts_supportable_tablet_peers` tends towards 1.0 as we approach the tablet peer limit for that TServer.

This PR adds a universe level alert with a warning when this ratio is more than 0.9 and severe if ratio is more than 1.0.

Test Plan:
Manually tested the alert:
* With a threshold of 0 on a universe - alert triggered (Expected)
* With default threshold on a universe - alert not triggered (Expected)

Reviewers: amalyshev, mlillibridge

Reviewed By: amalyshev

Subscribers: sanketh, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D34913