[PLAT-13644] Edit universe v2
Summary:
Implementation of edit universe in v2. Only a sub-set of the properties in create spec are editable. So creating an EditSpec for such objects. OTOH, for some objects, all the properties are editable. So using the create Spec for such objets in the edit request payload.

CRUD of ReadReplica clusters will be taken up in a subsequent diff.

Test Plan: Unit test.

Reviewers: dshubin, sanketh, #yba-api-review!, skurapati, dkumar

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D34782